+++ NEWS.md
@@ -1,5 +1,7 @@
 # dplyr (development version)
 
+* `nest_join()` now preserves the type of `y` (#6295).
+
 * Passing `...` to `across()` is now deprecated because the evaluation timing of
   `...` is ambiguous. Now instead of (e.g.) `across(a:b, mean, na.rm = TRUE)`
   you should write `across(a:b, ~ mean(.x, na.rm = TRUE))` (#6073).